refactor: save edit fees previous state if user clicks cancel
